sui_graphql_client

Module query_types

Modules

schema

Structs

ActiveValidatorsArgs

ActiveValidatorsQuery

Balance

BalanceArgs

BalanceQuery

Base64

BigInt

ChainIdentifierQuery

CheckpointArgs

CheckpointId

CheckpointQuery

CheckpointTotalTxQuery

CheckpointsArgs

CheckpointsQuery

CoinMetadata

The coin metadata associated with the given coin type.

CoinMetadataArgs

CoinMetadataQuery

DateTime

DefaultSuinsNameQuery

DefaultSuinsNameQueryArgs

DryRunArgs

DryRunQuery

DryRunResult

DynamicFieldArgs

DynamicFieldConnectionArgs

DynamicFieldName

DynamicFieldQuery

DynamicFieldsOwnerQuery

DynamicObjectFieldQuery

Epoch

EpochArgs

EpochQuery

EpochSummaryQuery

EpochValidator

EpochsArgs

EpochsQuery

Event

EventConnection

EventFilter

EventsQuery

EventsQueryArgs

ExecuteTransactionArgs

ExecuteTransactionQuery

ExecutionResult

GQLAddress

LatestPackageQuery

MoveFunction

MoveFunctionTypeParameter

MoveModule

MoveObject

MoveObjectContents

MovePackage

MovePackageConnection

MovePackageVersionFilter

MoveType

MoveValue

NormalizedMoveFunctionQuery

NormalizedMoveFunctionQueryArgs

NormalizedMoveModuleQuery

NormalizedMoveModuleQueryArgs

ObjectFilter

ObjectKey

ObjectQuery

ObjectQueryArgs

ObjectsQuery

ObjectsQueryArgs

OpenMoveType

Owner

PackageArgs

PackageByNameArgs

PackageByNameQuery

PackageCheckpointFilter

PackageQuery

PackageVersionsArgs

PackageVersionsQuery

PackagesQuery

PackagesQueryArgs

PageInfo

Information about pagination in a connection.

ProtocolConfigQuery

ProtocolConfigs

Information about the configuration of the protocol. Constants that control how the chain operates. These can only change during protocol upgrades which happen on epoch boundaries.

ProtocolVersionArgs

ResolveSuinsQuery

ResolveSuinsQueryArgs

ServiceConfig

ServiceConfigQuery

TransactionBlock

TransactionBlockArgs

TransactionBlockEffectsQuery

TransactionBlockQuery

TransactionBlockWithEffects

TransactionBlockWithEffectsQuery

TransactionBlocksEffectsQuery

TransactionBlocksQuery

TransactionBlocksQueryArgs

TransactionBlocksWithEffectsQuery

TransactionMetadata

TransactionsFilter

Validator

Represents a validator in the system.

ValidatorConnection

ValidatorSet

Enums

Feature

MoveAbility

MoveVisibility