sui_graphql_client::query_types

Struct Epoch

pub struct Epoch {

    pub epoch_id: u64,
    pub fund_inflow: Option<BigInt>,
    pub fund_outflow: Option<BigInt>,
    pub fund_size: Option<BigInt>,
    pub live_object_set_digest: Option<String>,
    pub net_inflow: Option<BigInt>,
    pub protocol_configs: Option<ProtocolConfigs>,
    pub reference_gas_price: Option<BigInt>,
    pub start_timestamp: DateTime,
    pub end_timestamp: Option<DateTime>,
    pub system_state_version: Option<u64>,
    pub total_checkpoints: Option<u64>,
    pub total_gas_fees: Option<BigInt>,
    pub total_stake_rewards: Option<BigInt>,
    pub total_stake_subsidies: Option<BigInt>,
    pub total_transactions: Option<u64>,
    pub validator_set: Option<ValidatorSet>,

}

Fields
epoch_id: u64

The epochโs id as a sequence number that starts at 0 and is incremented by one at every epoch change.

fund_inflow: Option<BigInt>

The storage fees paid for transactions executed during the epoch.

fund_outflow: Option<BigInt>

The storage fee rebates paid to users who deleted the data associated with past transactions.

fund_size: Option<BigInt>

The storage fund available in this epoch. This fund is used to redistribute storage fees from past transactions to future validators.

live_object_set_digest: Option<String>

A commitment by the committee at the end of epoch on the contents of the live object set at that time. This can be used to verify state snapshots.

net_inflow: Option<BigInt>

The difference between the fund inflow and outflow, representing the net amount of storage fees accumulated in this epoch.

protocol_configs: Option<ProtocolConfigs>

The epochโs corresponding protocol configuration, including the feature flags and the configuration options.

reference_gas_price: Option<BigInt>

The minimum gas price that a quorum of validators are guaranteed to sign a transaction for.

start_timestamp: DateTime

The epochโs starting timestamp.

end_timestamp: Option<DateTime>

The epochโs ending timestamp. Note that this is available only on epochs that have ended.

system_state_version: Option<u64>

The value of the version field of 0x5, the 0x3::sui::SuiSystemState object. This version changes whenever the fields contained in the system state object (held in a dynamic field attached to 0x5) change.

total_checkpoints: Option<u64>

The total number of checkpoints in this epoch.

total_gas_fees: Option<BigInt>

The total amount of gas fees (in MIST) that were paid in this epoch.

total_stake_rewards: Option<BigInt>

The total MIST rewarded as stake.

total_stake_subsidies: Option<BigInt>

The amount added to total gas fees to make up the total stake rewards.

total_transactions: Option<u64>

The total number of transaction in this epoch.

validator_set: Option<ValidatorSet>

Validator related properties. For active validators, see active_validators API.