sui_graphql_client::query_types
Struct Validator

pub struct Validator {

    pub apy: Option<i32>,
    pub address: GQLAddress,
    pub commission_rate: Option<i32>,
    pub credentials: Option<ValidatorCredentials>,
    pub description: Option<String>,
    pub exchange_rates_size: Option<u64>,
    pub gas_price: Option<BigInt>,
    pub name: Option<String>,
    pub image_url: Option<String>,
    pub next_epoch_commission_rate: Option<i32>,
    pub next_epoch_credentials: Option<ValidatorCredentials>,
    pub next_epoch_gas_price: Option<BigInt>,
    pub next_epoch_stake: Option<BigInt>,
    pub operation_cap: Option<MoveObject>,
    pub pending_pool_token_withdraw: Option<BigInt>,
    pub pending_stake: Option<BigInt>,
    pub pending_total_sui_withdraw: Option<BigInt>,
    pub pool_token_balance: Option<BigInt>,
    pub project_url: Option<String>,
    pub rewards_pool: Option<BigInt>,
    pub staking_pool_activation_epoch: Option<u64>,
    pub staking_pool_id: Address,
    pub staking_pool_sui_balance: Option<BigInt>,
    pub voting_power: Option<i32>,

}

Represents a validator in the system.

Fields
apy: Option<i32>

The APY of this validator in basis points. To get the APY in percentage, divide by 100.

address: GQLAddress

The validator’s address.

commission_rate: Option<i32>

The fee charged by the validator for staking services.

credentials: Option<ValidatorCredentials>

Validator’s credentials.

description: Option<String>

Validator’s description.

exchange_rates_size: Option<u64>

Number of exchange rates in the table.

gas_price: Option<BigInt>

The reference gas price for this epoch.

name: Option<String>

Validator’s name.

image_url: Option<String>

Validator’s url containing their custom image.

next_epoch_commission_rate: Option<i32>

The proposed next epoch fee for the validator’s staking services.

next_epoch_credentials: Option<ValidatorCredentials>

Validator’s credentials for the next epoch.

next_epoch_gas_price: Option<BigInt>

The validator’s gas price quote for the next epoch.

next_epoch_stake: Option<BigInt>

The total number of SUI tokens in this pool plus the pending stake amount for this epoch.

operation_cap: Option<MoveObject>

The validator’s current valid Cap object. Validators can delegate the operation ability to another address. The address holding this Cap object can then update the reference gas price and tallying rule on behalf of the validator.

pending_pool_token_withdraw: Option<BigInt>

Pending pool token withdrawn during the current epoch, emptied at epoch boundaries.

pending_stake: Option<BigInt>

Pending stake amount for this epoch.

pending_total_sui_withdraw: Option<BigInt>

Pending stake withdrawn during the current epoch, emptied at epoch boundaries.

pool_token_balance: Option<BigInt>

Total number of pool tokens issued by the pool.

project_url: Option<String>

Validator’s homepage URL.

rewards_pool: Option<BigInt>

The epoch stake rewards will be added here at the end of each epoch.

staking_pool_activation_epoch: Option<u64>

The epoch at which this pool became active.

staking_pool_id: Address

The ID of this validator’s 0x3::staking_pool::StakingPool.

staking_pool_sui_balance: Option<BigInt>

The total number of SUI tokens in this pool.

voting_power: Option<i32>

The voting power of this validator in basis points (e.g., 100 = 1% voting power).