Struct sui_graphql_client::Client

pub struct Client { /* private fields */ }

The GraphQL client for interacting with the Sui blockchain. By default, it uses the reqwest crate as the HTTP client.

Implementations

impl Client

pub fn new(server: &str) -> Result<Self, Error>

Create a new GraphQL client with the provided server address.

pub fn new_mainnet() -> Self

Create a new GraphQL client connected to the mainnet GraphQL server: {MAINNET_HOST}.

pub fn new_testnet() -> Self

Create a new GraphQL client connected to the testnet GraphQL server: {TESTNET_HOST}.

pub fn new_devnet() -> Self

Create a new GraphQL client connected to the devnet GraphQL server: {DEVNET_HOST}.

pub fn new_localhost() -> Self

Create a new GraphQL client connected to the localhost GraphQL server: {DEFAULT_LOCAL_HOST}.

pub fn set_rpc_server(&mut self, server: &str) -> Result<(), Error>

Set the server address for the GraphQL GraphQL client. It should be a valid URL with a host and optionally a port number.

pub async fn run_query<T, V>( &self, operation: &Operation<T, V>, ) -> Result<GraphQlResponse<T>>
where T: DeserializeOwned, V: Serialize,

Run a query on the GraphQL server and return the response. This method returns [cynic::GraphQlResponse] over the query type T, and it is intended to be used with custom queries.

pub async fn chain_id(&self) -> Result<String, Error>

Get the chain identifier.

pub async fn reference_gas_price( &self, epoch: Option<u64>, ) -> Result<Option<u64>, Error>

Get the reference gas price for the provided epoch or the last known one if no epoch is provided.

This will return Ok(None) if the epoch requested is not available in the GraphQL service (e.g., due to prunning).

pub async fn protocol_config( &self, version: Option<u64>, ) -> Result<Option<ProtocolConfigs>, Error>

Get the protocol configuration.

pub async fn service_config(&self) -> Result<ServiceConfig, Error>

Get the GraphQL service configuration, including complexity limits, read and mutation limits, supported versions, and others.

pub async fn active_validators( &self, epoch: Option<u64>, after: Option<String>, before: Option<String>, first: Option<i32>, last: Option<i32>, ) -> Result<Option<Page<Validator>>, Error>

Get the list of active validators for the provided epoch, including related metadata. If no epoch is provided, it will return the active validators for the current epoch.

pub async fn balance( &self, address: Address, coin_type: Option<&str>, ) -> Result<Option<u128>, Error>

Get the balance of all the coins owned by address for the provided coin type. Coin type will default to 0x2::coin::Coin<0x2::sui::SUI> if not provided.

pub async fn coins( &self, owner: Address, after: Option<&str>, before: Option<&str>, first: Option<i32>, last: Option<i32>, coin_type: Option<&str>, ) -> Result<Option<Page<Coin<'_>>>, Error>

Get the list of coins for the specified address.

If coin_type is not provided, it will default to 0x2::coin::Coin, which will return all coins. For SUI coin, pass in the coin type: 0x2::coin::Coin<0x2::sui::SUI>.

pub fn coins_stream<'a>( &'a self, owner: Address, coin_type: Option<&'a str>, ) -> Pin<Box<dyn Stream<Item = Result<Coin<'_>, Error>> + 'a>>

Stream of coins for the specified address and coin type.

pub async fn coin_metadata( &self, coin_type: &str, ) -> Result<Option<CoinMetadata>, Error>

Get the coin metadata for the coin type.

pub async fn total_supply(&self, coin_type: &str) -> Result<Option<u64>, Error>

Get total supply for the coin type.

pub async fn checkpoint( &self, digest: Option<String>, seq_num: Option<u64>, ) -> Result<Option<CheckpointSummary>, Error>

Get the CheckpointSummary for a given checkpoint digest or checkpoint id. If none is provided, it will use the last known checkpoint id.

pub async fn checkpoints( &self, after: Option<&str>, before: Option<&str>, first: Option<i32>, last: Option<i32>, ) -> Result<Option<Page<CheckpointSummary>>, Error>

Get a page of CheckpointSummary for the provided parameters.

pub async fn latest_checkpoint_sequence_number( &self, ) -> Result<Option<CheckpointSequenceNumber>, Error>

Return the sequence number of the latest checkpoint that has been executed.

pub async fn epoch_total_checkpoints( &self, epoch: Option<u64>, ) -> Result<Option<u64>, Error>

Return the number of checkpoints in this epoch. This will return Ok(None) if the epoch requested is not available in the GraphQL service (e.g., due to prunning).

pub async fn epoch_total_transaction_blocks( &self, epoch: Option<u64>, ) -> Result<Option<u64>, Error>

Return the number of transaction blocks in this epoch. This will return Ok(None) if the epoch requested is not available in the GraphQL service (e.g., due to prunning).

pub async fn events( &self, filter: Option<EventFilter>, after: Option<String>, before: Option<String>, first: Option<i32>, last: Option<i32>, ) -> Result<Option<Page<Event>>, Error>

pub async fn object( &self, address: Address, version: Option<u64>, ) -> Result<Option<Object>, Error>

Return an object based on the provided Address.

If the object does not exist (e.g., due to prunning), this will return Ok(None). Similarly, if this is not an object but an address, it will return Ok(None).

pub async fn objects( &self, after: Option<&str>, before: Option<&str>, filter: Option<ObjectFilter<'_>>, first: Option<i32>, last: Option<i32>, ) -> Result<Option<Page<Object>>, Error>

Return a page of objects based on the provided parameters.

Use this function together with the ObjectFilter::owner to get the objects owned by an address.

Example

let filter = ObjectFilter {
    type_: None,
    owner: Some(Address::from_str("test").unwrap().into()),
    object_ids: None,
    object_keys: None,
};

let owned_objects = client.objects(None, None, Some(filter), None, None).await;

pub async fn object_bcs( &self, object_id: Address, ) -> Result<Option<Vec<u8>>, Error>

Return the object’s bcs content Vec<u8> based on the provided Address.

pub async fn dry_run_tx( &self, tx: &Transaction, skip_checks: Option<bool>, ) -> Result<DryRunResult, Error>

Dry run a Transaction and return the transaction effects and dry run error (if any).

skipChecks optional flag disables the usual verification checks that prevent access to objects that are owned by addresses other than the sender, and calling non-public, non-entry functions, and some other checks. Defaults to false.

pub async fn dry_run_tx_kind( &self, tx_kind: &TransactionKind, skip_checks: Option<bool>, tx_meta: TransactionMetadata, ) -> Result<DryRunResult, Error>

Dry run a TransactionKind and return the transaction effects and dry run error (if any).

skipChecks optional flag disables the usual verification checks that prevent access to objects that are owned by addresses other than the sender, and calling non-public, non-entry functions, and some other checks. Defaults to false.

tx_meta is the transaction metadata.

pub async fn transaction( &self, digest: &str, ) -> Result<Option<SignedTransaction>, Error>

Get a transaction by its digest.

pub async fn transactions<'a>( &self, after: Option<&str>, before: Option<&str>, first: Option<i32>, last: Option<i32>, filter: Option<TransactionsFilter<'a>>, ) -> Result<Option<Page<SignedTransaction>>, Error>

Get a page of transactions based on the provided filters.

pub async fn execute_tx( &self, signatures: Vec<UserSignature>, tx: &Transaction, ) -> Result<Option<TransactionEffects>, Error>

Execute a transaction.